adding things to remote
