+++ src/status.js
@@ -1,0 +1,12 @@
+export function saveLocal(list) {
+  window.localStorage.setItem('localTasks', JSON.stringify(list));
+}
+
+export function status(elem, list) {
+  list.forEach((task) => {
+    if (task === elem) {
+      task.isCompleted = !task.isCompleted;
+    }
+  });
+  saveLocal(list);
+}
